Fix hostname condition in how-to page
On line 219, we can read the following:

```
{{- if ne .chezmoi.hostname "work-laptop" }}
    .work # only manage .work on work-laptop
    {{- end }}
```
The comment in the condition hints at the body of the condition will be evaluated if this is the hostname is `work-laptop`. However, the condition uses negation instead of equality.

I suggest to update it to:

```
{{- if eq .chezmoi.hostname "work-laptop" }}
```
so the condition reflects the intention and also matches the same condition on line 176.