docs(REFERENCE): clarify .chezmoiignore patterns
Docs incorrectly state that the pattern

    backups/**

in a `.chezmoiignore` file would ignore `backups` and all its
contents recursively; the pattern ignores all the contents of backups
but not the directory itself. (As a consequence, `chezmoi add` would add
`backups` with a `.keep` file in it, but not any of the files or folders
in `backups` itself.)

This behavior comes from the `doublestar` library, which has a test
ensuring that `**` patterns work this way, so this is presumably the
intended behavior.