Default to ~/.local/bin rather than ~/bin
The directory `~/bin` is an antiquated 20 year old social convention that came out of need and a failure on the part of the Linux Filesystem Hierarchy Standard (FHS) to state anything on the matter. The new replacement for that is standardized by the Systemd File Hierarchy Specification; see `man file-hierarchy` for more information. It states `~/.local/bin/` is the right place for user-installed binary files.