fix: Fix op signin for 1Password CLI 2.0.0
The release of op 2.0.0 introduced a breaking change to `op signin` so
that the previous version invocation `op signin account` no longer
works. Now, it is required to use the option `--account`, as in `op
signin --account account`.

Note that there is likely another issue hidden by this when not using
biometric authentication for `op`, because the session environment
variable is no longer `OP_SESSION_<account>` but
`OP_SESSION_<accountUUID>`. If the account is specified as the account
UUID, everything will probably work. I’m using biometric authentication,
so I cannot easily verify this.

Also note that in `op` v1, you could specify `--vault ''`; this is no
longer legal.

https://1password.community/discussion/128083/problems-with-op-v2-when-integrating-with-chezmoi#latest