docs: Add clarification about all lowercase custom data fields
Hi, 

I'm new to chezmoi and started to look into templates. I've spent quite some time to figure out why I could not get custom variables in the `data` section of the config to work until I realized that I've stumbled upon #463.

I propose to add a hint or disclaimer to the documentation here to make it easier for newcomers to spot this pitfall (and avoid it).

Please let me know if I need to change or update anything.