docs: Add example of extracting selected files from external archive
This is an add on to the previous fix for #2056 which adds an example of
how to use `input` filters in `.chezmoiexternal.toml` to the
user guide (`docs/user-guide/include-files-from-elsewhere.md`).

The commit also includes an update to
`docs/reference/special-files-and-directories/chezmoiexternal-format.md`
which uses the new simplified `input` filter syntax and makes a minor
change to the comment about excluded directory trees in
`pkg/chezmoi/sourcestate.go`.