feat: Improve performance of the externals cache
Previously, chezmoi stored external cache entries in gzip'ed JSON files,
with the JSON containing the URL, time of download, and the hex encoded
contents. This meant that reading the contents cache required gzip
decompression, JSON unmarshaling, and hex decoding.

This commit changes the cache format to store the raw content with the
time of download stored in the modified time of the file, eliminating
the gzip decompression, JSON unmarshalling, and hex decoding steps.

A side effect is that old cache entries will not be read or cleaned up.

To remove old cache entries, run:

    rm ~/.cache/chezmoi/external/*.json.gz