Expose some chezmoi config in env vars
I've been playing with some of the runtime scripts in my configuration
and realized that I could change most of them—if not all—from
`run_*.sh.tmpl` to just regular shell scripts (so that I can run them
without clearing state for things that occasionally get reset by the
OS) if Chezmoi exported some of its configuration as environment
variables.

The template values that I care about most are `.chezmoi.homeDir` and
`.chezmoi.sourceDir`, but other values may be useful as well, so I have
set the following values when the configuration object is constructed:

- `CHEZMOI_ARCH`
- `CHEZMOI_CACHE_DIR`
- `CHEZMOI_CONFIG_FILE`
- `CHEZMOI_HOME_DIR`
- `CHEZMOI_OS`
- `CHEZMOI_SOURCE_DIR`
- `CHEZMOI_VERSION`
- `CHEZMOI_WORKING_TREE`

In general, my preference would be to set these *immediately* before
`realSystem.RunCommand` using `exec.Command.Env`, but that would require
changes to a number of other APIs (the config is not passed to
`RunCommand`) and be slower in any case as the variables would need to
be set on each script execution.

This *should* work on Windows (it is guaranteed to work on any Unix-like
OS), but I don't have a Windows system to verify. I'm uncertain what
sort of tests could be run or updated to show this working, but "it
works on my system".