fix: Construct `templateDataMap` manually
Closes #2502

Round-tripping the template data via JSON caused a data regression such
that values that were `map[string]any{nil}` (the default value) are
turned into `null` in JSON and therefore `nil` in a template.
Previously, the value of `map[string]any{nil}` would be rendered as
`{}`.

Manual construction of the map restores the old functionality. For
chezmoi version 3, we may want to consider *restoring* the behaviour
from `v2.25.0` (`null`), but this is a regression version version 2.

```console
# before this fix
$ go run . execute-template '{{ .chezmoi.osRelease | toJson }}'
null%
$ go run . execute-template '{{ .chezmoi.osRelease | toJson }}'
{}%
```