fix: Eliminate [DEFAULT] section for fromIni map
Closes #2768

In `fromIni`, the `iniFileToMap` function converts top-level keys (the
so-called "default" section) to being members of a `DEFAULT` map key,
which results in different data than expected. Specifically,
roundtripping fails such that `{{ . | fromIni | toIni }}` does not
produce the same output as input.

This change results in code duplication, but prevents the unnecessary
creation of a new `map[string]any` or a second loop for merging.