fix: Prevent AppleDouble file creation in .txtar
Related to #3220 and fixes it for tests run on macOS (but does not fix
the underlying issue).

Any time we create an archive with `exec tar c…` in a `txtar` test file,
particularly where the golden contents will be compared, we need to have
a separate command for `darwin` platforms:

```txtar
[!darwin] tar cfz archive.tar.gz archive
[darwin] tar cfz archive.tar.gz --no-mac-metadata archive
```

Otherwise, the actual output will be incorrect compared to the golden
version on macOS because of the presence of
[AppleDouble](https://en.wikipedia.org/wiki/AppleSingle_and_AppleDouble_formats)
files. As a developer on macOS, this is helpful because it will give me
an improved test cycle before submitting for CI.