feat: Ignore AppleDouble files in externals
This is somewhere between a fix and a feature because it is a potential
bug if an archive contains AppleDouble data, as such data would
automatically be written out as parallel `._FILE` files and not
reintegrated.

With this change, we are still not reintegrating the data, but we are
explicitly *ignoring* the data unless
`EXTERNAL.archive.extractAppleDoubleFiles = true`. There are tools such
as `https://github.com/SiviourBla/appledouble` that can theoretically
reintegrate this using `copyfile(3)`.

This checks that the extracted file name will be `._*` before trying to
determine whether the file is an AppleDouble file.

- This also adds `/.venv` directories to the `lint-whitespace` ignore
  list. I am exploring the use of Python venvs for using mkdocs.

Resolves: #3220
Closes: #3221