fix: keep initFuncs when using 'includeTemplate' in config
When trying to exec 'chezmoi init' or 'chezmoi apply --init'  with
.chezmoi.toml.tmpl including another file through 'includeTemplate',
the initTemplateFuncs are only available in the first file,
but not in the included template.

This is because the initial execution operates on a copy of c.templateFuncs,
which wouldn't get passed on to includeTemplateTemplateFunc.

With this change, initFunctions will be added directly to c.templateFuncs
and then removed after the config has been generated.

Note: Due to a separate implementation, 'chezmoi execute-template --init'
is not affected by this bug.