docs: Use `uv` for managing Python and dependencies
Python packaging is in flux, but the old form of requirements
(`requirements.txt`) is no longer recommended. The use of a more modern
alternative (such as Poetry, PDM, Rye, or uv) is recommended as they
also provide dependency lock files so that downstream dependencies are
not updated on installation.

For a project that is not written in Python, `uv` looks likely to be the
best choice because it can also manage the Python installation, using a
suitable existing Python installation if present.

As `uv` does not currently support running tasks (unlike Poetry and
pdm), we have added `taskipy` with useful shorthand tasks defined for
this purpose:

| name        | command                                      |
| ----------- | -------------------------------------------- |
| build-docs  | mkdocs build -f assets/chezmoi.io/mkdocs.yml |
| serve-docs  | mkdocs serve -f assets/chezmoi.io/mkdocs.yml |
| deploy-docs | cd assets/chezmoi.io && mkdocs gh-deploy     |
| lint        | ruff check                                   |
| format      | ruff format                                  |
| pycheck     | task lint && task format --diff              |
| pyfix       | task lint --fix && task format               |
| format-yaml | make format-yaml                             |

These are run with `uv run task <name>`. Note that `make format-yaml`
is the preferred way to format YAML documents, and it runs the
`format-yaml.py` script directly.`