chore: Improved `$FORMAT` and `.chezmoidata` docs
There have been an increased number of issues filed trying to treat
`.chezmoidata.$FORMAT` and `.chezmoidata/` as templates. This should
resolve this and make it easier to handle.

Additionally, as I'm preparing to use `.chezmoidata` files more
personally, I was mildly *surprised* by the fact that the path to the
`.chezmoidata.$FORMAT` or the filenames in `.chezmoidata/` directories
were not treated as partial key prefixes the way that such data files
are handled in Jekyll, Hugo, etc. It is worth documenting the data merge
behaviour to reduce end-user surprise.