chore: Upgrade Mermaid version
Force `mermaid2` to the latest version (11.4.1).

This resulted in a minor bug on rendering where the palette numbering
was apparently changed from 1 (light) and 2 (dark) to 0 (light) and
1 (dark). Updated the `refresh_on_toggle_dark_light.js` script to work
with either scenario. `document.getElementById("__palette_0")` will
return `null` if using 1 indexing, so we swap the dark palette switcher
with the light palette switcher and set the dark palette switcher to
`__palette_2`.