fix: Set .chezmoi.sourceFile in execute-template command
If a template file's path is in the chezmoi source directory, when
running `chezmoi execute-template -f FILENAME`, set
`.chezmoi.sourceFile` to the same source directory-relative path that a
template executing via `chezmoi apply` would have.

If a template's path is *not* in the source directory, then set
`.chezmoi.sourceFile` to the filename exactly as it is given on the
command line.

Previously, all template files executed via `chezmoi execute-template`
would get `.chezmoi.sourceFile` set to the string `"argN"`, where N is
the argument position on the command line, which breaks testing any
templates that use `sourceFile`.

Signed-off-by: Ross Williams <ross@ross-williams.net>